Extract EmailValidationForm component
